Update mpi_to_openshmem.c
Added two helper functions for MPI_Bcast.  One to copy data from one buffer to a symmetric buffer (or the other way). One to calculate the offset in the symmetric buffer, so we do not over write data...

MPI_Bcast allows users to use their local memory or shmalloc memory.